# api/openapi.yaml
openapi: 3.1.0
info:
  title: 艦隊分析 応募者管理API
  version: 0.3.0
  description: 艦隊分析への応募を管理するAPI。仕様は docs/spec.md を参照。
servers:
  - url: https://api.example.com
    description: 本番（ドメインは未定）
security:
  - cognito: []
...
paths:
...
  /stream/current:
    get:
      tags: [stream]
      operationId: getStreamView
      summary: 配信用画面の表示内容（いま分析中の1人）
      responses:
        '200':
          description: 表示内容。分析中の人がいなければ current は null
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /stream/next:
    post:
      tags: [stream]
      operationId: advanceStream
      summary: 次の人へ進む（配信者・運営）
      description: |
        いま分析中の人を「分析済み」にし、「次に分析する順」の先頭の人を「分析中」にする。
        分析予定の人はそのまま選び、未着手の人は重複・条件外の印があれば飛ばす。
        抽選を使う設定のときは未着手の人を選ばない（抽選で選ぶため）。
        次の人がいなければ、分析済みにするだけで current は null になる。
      responses:
        '200':
          description: 進めた後の表示内容
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /stream/operators:
    get:
      tags: [stream]
      operationId: listStreamOperators
      summary: 関係者ごとの、配信の操作の許可（配信者・運営）
      responses:
        '200':
          description: 関係者の一覧と、それぞれ許可しているか
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StreamOperator'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /stream/operators/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
    put:
      tags: [stream]
      operationId: grantStreamOperator
      summary: 関係者に配信の操作を許可する（配信者・運営）
      responses:
        '204':
          description: 許可した
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [stream]
      operationId: revokeStreamOperator
      summary: 関係者への配信の操作の許可を取り消す（配信者・運営）
      responses:
        '204':
          description: 取り消した
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    cognito:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: CognitoのIDトークン（手元では /dev/login の仮のトークン）。グループ（admin/streamer/staff）で権限を判定
    formKey:
      type: apiKey
      in: header
      name: X-Form-Key
      description: Apps Scriptだけが持つ秘密キー
...
  responses:
    BadRequest:
      description: 入力が不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: 認証が必要（トークンがない・期限切れ・正しくない）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: 権限がない（二段階認証をまだ設定していない配信者・運営は、どの操作もこれになる）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: 見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required: [code, message]
...
    StreamOperator:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: string
        allowed:
          type: boolean
      required: [userId, displayName, allowed]
...
    Answers:
      type: object
      description: フォーム回答。キーは設定の項目コード。版ごとに項目が変わるため自由形式
      additionalProperties: true
...
    StreamView:
      type: object
      properties:
        current:
          oneOf:
            - $ref: '#/components/schemas/StreamApplicant'
            - type: 'null'
      required: [current]
    StreamApplicant:
      type: object
      description: 配信に映す情報のみ。XのIDと課金額は含めない
      properties:
        displayName:
          type: string
          description: 匿名希望なら「匿名提督」など
        simulatorUrl:
          type: string
          format: uri
        answers:
          $ref: '#/components/schemas/Answers'
        previous:
          oneOf:
            - $ref: '#/components/schemas/Answers'
            - type: 'null'
          description: 前回分析した応募（同じXのIDで分析済みのうち一番新しいもの）の回答（比較用）
      required: [displayName, simulatorUrl, answers, previous]
...